Package publishing settings
